Add files via upload
This program was developed to control the hardware of an automated arduino car, examples being the motors and an ultrasonic sensor of the car. It was also programmed to be controlled with an app that used Google's Speech-to-Text API to control the actions of the car (eg. go forwards, turn left).